# crimson: split ObjectDataBlock when overwriting