This PR introduces a straightforward strategy to reduce write amplification.
The design and changes can be found in the commit message and comments.

# Benchmark
The benchmark of splitting extents contains two major aspects:
1. write to new object. In such cases, no extra read and write occurs, it can get the overhead of maintaining the data structure and the procedure of splitting.
2. overwrite, the length of original extent is 4KiB and greater. When the length is greater than 4KiB, splitting can reduce the read and write amplification so that performance could be better.

for 1. create a new rbd image and test with fio;
for 2. create a rbd image and sequential write, then random write to the full image.

start command:

```
MGR=1 MON=1 OSD=1 MDS=0 RGW=0 ../src/vstart.sh -n -x --without-dashboard --seastore --crimson --redirect-output --seastore-devs /dev/nvme1n1
```

## 1. random write to empty image

<details>
<summary>fio config</summary>

```
[global]
ioengine=rbd
clientname=admin
pool=test_pool
rbdname=test_img
direct=1
sync=1
time_based=1
runtime=600
randseed=1259
[randwrite-4k]
rw=randwrite
bs=4K
numjobs=1
iodepth=768
random_distribution=zoned:90/10:10/90
```

</details>

Result:
|        | IOPS | BW        |
| --- | --- | --- |
| master | 2637 | 10.3MiB/s |
| split  | 3037 | 11.9MiB/s |

## 2. random write to full image

### 2.1. the length of extent in image is 4KiB

<details>
<summary> sequential write with fio </summary>

```
[global]
ioengine=rbd
clientname=admin
pool=test_pool
rbdname=test_img
direct=1
sync=1
randseed=1259
[write-4k]
rw=write
bs=4K
numjobs=1
iodepth=768
random_distribution=zoned:90/10:10/90
```

</details>

|        | IOPS  | BW        |
| --- | --- | --- |
| master | 20.8k | 81.3MiB/s |
| split  | 20.3k | 79.2MiB/s |

After the image is full, start random writes

<details>
<summary> random write with fio </summary>

```
[global]
ioengine=rbd
clientname=admin
pool=test_pool
rbdname=test_img
direct=1
sync=1
time_based=1
runtime=600
randseed=1259
[randwrite-4k]
rw=randwrite
bs=4K
numjobs=1
iodepth=768
random_distribution=zoned:90/10:10/90
```

</details>

|        | IOPS | BW        |
| --- | --- | --- |
| master |  141 | 567KiB/s  |
| split  | 2307 | 9231KiB/s |

### 2.2. the length of extent in image is 64KiB

<details>
<summary> sequential write with fio </summary>

```
[global]
ioengine=rbd
clientname=admin
pool=test_pool
rbdname=test_img
direct=1
sync=1
randseed=1259
[write-4k]
rw=write
bs=64K
numjobs=1
iodepth=768
random_distribution=zoned:90/10:10/90
```

</details>

|        | IOPS | BW       |
| --- | --- | --- |
| master | 8530 | 533MiB/s |
| split  | 7484 | 468MiB/s |

After the image is full, start random writes

<details>
<summary> random write with fio </summary>

```
[global]
ioengine=rbd
clientname=admin
pool=test_pool
rbdname=test_img
direct=1
sync=1
time_based=1
runtime=600
randseed=1259
[randwrite-4k]
rw=randwrite
bs=4K
numjobs=1
iodepth=768
random_distribution=zoned:90/10:10/90
```

</details>

|        | IOPS | BW        |
| --- | --- | --- |
| master |   77 | 311KiB/s  |
| split  | 2671 | 10.4MiB/s |



## 3. summary
The results are annoying. The performance of random writes to a full image is much
worse than expected that the difference of 2.1 test could be insignificant. There may be other factors present.

But the results of splitting are as expected:
writing to empty image has best performance;
writing to 64KiB extent is second;
writing to 4KiB extent is last. 

<!--
  - Please give your pull request a title like

      [component]: [short description]

  - Please use this format for each git commit message:

      [component]: [short description]

      [A longer multiline description]

      Fixes: [ticket URL on tracker.ceph.com, create one if necessary]
      Signed-off-by: [Your Name] <[your email]>

    For examples, use "git log".

  - The Signed-off-by line in every git commit is important; see <span class="x x-first x-last">[Submitting Patches to Ceph](https://github.com/ceph/ceph/blob/master/</span>SubmittingPatches.rst<span class="x x-first x-last">)</span>
-->

## Checklist
- Tracker (select at least one)
  - [ ] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [ ] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [ ] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [ ] No doc update is appropriate
- Tests (select at least one)
  - [ ] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [ ] No tests

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
</details>