@athanatos 2022-04-04
This PR introduces split in a way that cuts across existing abstraction barriers and increases complexity greatly.  Have you considered any other ways to do this that fit better within existing abstractions?