@athanatos 2022-04-04
Is the problem here that the extent_info_t within the records can't be modified after write?  The backref change should fix that, would that remove the need for the extent retirement complexity and split_context_t?