@zhscn 2022-04-06
> Have you considered any other ways to do this that fit better within existing abstractions?

Yes and no. 1) For yes, I did design another solution that introduces a new "write context" layer in which we first do splits without the help of Cache/LBAManager and only calls Cache/LBAManager's methods when leaving the ObjectDataHandler, this design can avoid making changes to "Cache::get_extent()", "TransactionManager::dec_ref()" and "Cache::retire_extent_addr()". 2) For no, current abstraction can not handle with the cases that one transaction contains multiple split operation on one extent, which makes the changes of existing abstractions unavoidable. Details are explained later.

> Is the problem here that the extent_info_t within the records can't be modified after write? The backref change should fix that, would that remove the need for the extent retirement complexity and split_context_t?

Not exactly, backrefs only affects rewrite process which is not leading to changes in components other than SegmentCleaner.

# Current design

Here are some design trade-offs needs to be considered:

Actually splitting process affects 3 aspects:

1. ObjectDataHandler, the construction phase of a transaction. Currently all split operations occur at this period. 

First problem: whether the remaining pieces of an ObjectDataBlock should exist in the form of CachedExtent after splitting and before completing transaction. 

```
                                  |
               Transaction begin  |
               ObjectDataHandler  |
                                  |
                                  |                |******| incoming data (not a CachedExtent)
                                  |
                                  |   |-------------------| original extent (ObjectDataBlock with paddr1, real and not relative)
                                  |
                   split extent   |
         remove original extent   |
alloc extent for incoming  data   |                |******| ObjectDataBlock with paddr2, relative, in write_set
                                  |
                                  |   |------------|       original-left piece of the extent
                                  |                        (is this a CachedExtent? What's its paddr, paddr1 or a new relative paddr3?)
                                  |
                                  |
                                  v
```

For the original-left piece, 1) if it is a CachedExtent with the same physical address of original ObjectDataBlock paddr1, then Cache::get_extent can not work correctly since its extent is indexed only by paddr_t, so we need a new structure to hold the original-left piece; 2) if this piece has a new releative paddr3, then there must be a data structure in Transaction to maintain the mapping from new relative paddr3 to original real paddr1.

Second problem: how to handle operations on "split extents"? 

```
                                  |
               Transaction begin  |
               ObjectDataHandler  |
                                  |
                                  |               |******| incoming data (not a CachedExtent)
                                  |
                                  |  |-------------------| original extent (ObjectDataBlock with paddr1, real and not relative)
                                  |
                   split extent   |
         remove original extent   |
alloc extent for incoming  data   |               |******| ObjectDataBlock with paddr2, relative, in write_set
                                  |
                                  |  |------------|       original-left piece of the extent
                                  |
                                  |
                                  |
               next operation     |
                                  |      +---------------  result of previous split(we call it "split extent")
                                  |      |
                                  |      |           +---  incoming data of prev op
                                  |      v           v     ObjectDataBlock with relative paddr
                                  |  |------------|******|
                                  |
                                  |   |****|               new incoming data, offset and
                                  |                        length are not block size aligned
                                  v
```

With current mechanism, every step of handling transaction operation works as follows:  
- get pin from LBA by laddr_t  
- get extent from Cache by paddr_t.  

If the pin returned from LBA is the "split pin" of previous operation, then all operations on this pin needs to be transacted into operations on the original extent.

For example, remove a "split extent": Transaction::dec_ref involves Cache::retire_extent_addr, but a "split extent" is the part of original extent which has been retired at previous operation. So removing a "split extent" should not call Cache::retire_extent_addr and only remove entry from LBA tree.

There must exist a mechanism to distinguish whether a range of (laddr_t,length) corresponds to a "split extent" of previous operation. 

Given these two problems, a "split_context_t" is necessary to maintain the information about the split extents.

What's more, this PR doesn't treat "split extent" as a real CachedExtent, so allocating split extents is only allocating a new entry on LBA tree, not allocating CachedExtent from Cache.

2. extent retirement.

Follow the part 1, since a split operation involves removing old extents and inserting new extents, old extents are always in Transaction::retired_set. In case of partially overwrite, old extents should not be completely retired, as the remaining pieces of them are still live. So when retiring an ObjectDataBlock, we also needs a structure to find out retired ranges correctly, which is also the purpose of split_context_t.

3. rewrite extent. 

Currently checking a extent(ObjectDataBlock) are alive calls LBAManager::get_pins. The more efficient way will be calling backref tree in the future.

# Write Context

Of course this design causes several break changes: 1) Cache::get_extent can return retired ObjectDataBlock; 2) complex parameters of Transaction::dec_ref and Cache::retire_extent_addr; 3) complex retire policy 4) and some others.

The root cause is that for each transaction step, we do the following sequence:  
1. load touched extents  
2. do splits  
3. alloc "split extents" and "new extents"
And successive transaction steps needs to deal with the result of the previous transaction steo.

With the "write context" approach, we attach each transaction with a write_context_t, and cache all steps' results in it, so the workflow changes to the following:
1. load all extents that are touched by all transaction ops
2. do steps on top of the write_context_t, some of which mean involve extents splitting
3. do all retires/allocations

Hopefully, with the "write context" approach, the first two break changes above, that's "Cache::get_extent can return retired ObjectDataBlock" and "complex parameters of Transaction::dec_ref and Cache::retire_extent_addr", can be avoided. And the third issue of "complex retiring policy" can be simplified.